Documentation

Documentation

Functions

Index

Functions

Converting

ChaoticMessages.convert_message_to_samples — Method

convert_message_to_samples(
    message::String;
    add_noise::Bool = false,
    std::Float64 = 1.0,
)

Arguments

message::String : filepath to the wav file to convert to samples

add_noise::Bool : if true, add Gaussian noise of standard deviation std and none otherwise

std::Float64 : standard deviation of Gaussian noise if add_noise is true

Returns

func_message_unencrypted : waveform that is a function of time

num_samples : number of samples

sampling_rate : sampling rate

Convert a message (.wav file) which is identified as the string of the file path and convert it into samples. If add_noise is true, then noise is added to the samples. The noise is generated from a Gaussian distribution with mean 0 and standard deviation std`.

ChaoticMessages.convert_samples_to_message — Method

convert_samples_to_message(
    samples,
    sampling_rate,
    num_of_samples::Int64,
    name_of_file::String,
)

Arguments

samples : function that produce the samples

sampling_rate : sampling rate of the resulting file

num_of_samples::Int64 : number of samples of the resulting file

name_of_file::String : name of the resulting file

Convert samples into a wav file and save it.

ChaoticMessages.binary_to_bmessage — Method

binary_to_bmessage(
    s::String;
    time_length::Float64 = 2.0,
    b_zero::Float64 = 4.0,
    b_one::Float64 = 4.4,
)

Arguments

s::String : binary string of 0's and 1's

time_length::Float64 = 2.0 : time between each binary digit

b_zero::Float64 = 4.0 : value of the function when the binary digit is zero

b_one::Float64 = 4.4 : value of the function when the binary digit is one

Return a function that map the time $t$ to floor(t/time_length)th position of the binary string. The binary digits in the string is counted starting with 0. If t < 0 or t >= time_length * length(binary_arr), then the function evaluate to b_zero.

Encrypting

ChaoticMessages.create_secret_message — Method

create_secret_message(
    u0::Vector{Float64},
    p::Vector{Float64},
    tspan::Union{Tuple{Float64,Float64},Vector{Float64}},
    message_unencrypted;
    binary::Bool = false,
    scale::Float64 = 1e-5,
    abstol::Float64 = 1e-11,
    reltol::Float64 = 1e-11,
)

Arguments

u0::Vector{Float64}: initial condition of the Lorenz system

p::Vector{Float64}: parameters of the Lorenz system

tspan::Union{Tuple{Float64, Float64}, Vector{Float64}} : length of the message

message_unencrypted : function that produces the message that is to be encrpyted

binary::Bool = false : indicates whether the message being encrpyted is binary or not

scale::Float64 = 1e-5 : scaling factor for the secret message

abstol::Float64 = 1e-11 : absolute tolerance for the ODE solver

reltol::Float64 = 1e-11 : relative tolerance for the ODE solver

Encrpyt the message using a paramterized version of the Lorenz system (see lorenz_transmitter! and lorenz_transmitter_binary!).

ChaoticMessages.decrypt_secret_message — Method

decrypt_secret_message(
    u0::Vector{Float64},
    p::Vector{Float64},
    tspan::Union{Tuple{Float64,Float64},Vector{Float64}},
    secret_message;
    binary::Bool = false,
    scale::Float64 = 1e5,
    abstol::Float64 = 1e-11,
    reltol::Float64 = 1e-11,
)

Arguments

u0::Vector{Float64}: initial condition of the variation of the Lorenz system

p::Vector{Float64}: parameters of the variation of the Lorenz system

tspan::Union{Tuple{Float64, Float64}, Vector{Float64}} : length of the message

secret_message : function that produces the secret message that is to be decrypted

binary::Bool = false : indicates whether the message being decrypted is binary or not

scale::Float64 = 1e-5 : scaling factor for the recovered message

abstol::Float64 = 1e-11 : absolute tolerance for the ODE solver

reltol::Float64 = 1e-11 : relative tolerance for the ODE solver

Decrypt the message using a paramterized version of the Lorenz system (lorenz_receiver!).

Chaotic Systems

ChaoticMessages.lorenz_transmitter! — Method

lorenz_transmitter!(du, u, p, t)

Define the parameterized Lorenz system which is used to transmit messages.

The system of differential equations is -

$\dot{x_T} = \sigma (y_T - x_T),$

$\dot{y_T} = rx_T - y_T - 20 x_T z_T,$

$\dot{z_T} = 5x_T y_T - b z_T.$

ChaoticMessages.lorenz_receiver! — Method

lorenz_receiver!(secret_message)

Define the parameterized Lorenz system which is used to receive the encrpyted message.

The system of differential equations is -

$\dot{x_R} = \sigma (y_R - x_R),$

$\dot{y_R} = r \cdot m(t) - y_R - 20 (m(t) \cdot z_R),$

$\dot{z_R} = 5 (m(t) \cdot y_R) - b \cdot z_R.$

ChaoticMessages.lorenz_transmitter_binary! — Method

lorenz_transmitter_binary!(du, u, p, t)

Define the parameterized Lorenz system which is used to transmit binary messages.

The system of differential equations is -

$\dot{x_T} = \sigma (y_T - x_T),$

$\dot{y_T} = rx_T - y_T - 20 x_T z_T,$

$\dot{z_T} = 5x_T y_T - b(t) z_T.$

Notice that the only difference between lorenz_transmitter! and lorenz_transmitter_binary! is that b is a function of time.

Index

ChaoticMessages.binary_to_bmessage

ChaoticMessages.convert_message_to_samples

ChaoticMessages.convert_samples_to_message

ChaoticMessages.create_secret_message

ChaoticMessages.decrypt_secret_message

ChaoticMessages.lorenz_receiver!

ChaoticMessages.lorenz_transmitter!

ChaoticMessages.lorenz_transmitter_binary!